feat: add zustand & struture clean
