+++ app/constants/cities.js
@@ -1,0 +1,9 @@
+export const cities = [
+  'Ankara',
+  'Istanbul',
+  'Izmir',
+  'Antalya',
+  'Bursa',
+  'Edirne',
+  'Gaziantep'
+];
