+++ app/constants/countries.js
@@ -1,0 +1,19 @@
+export const countries = [
+  { value: 'France', label: 'Fransa', flag: '🇫🇷' },
+  { value: 'Netherlands', label: 'Hollanda', flag: '🇳🇱' },
+  { value: 'Ireland', label: 'İrlanda', flag: '🇮🇪' },
+  { value: 'Malta', label: 'Malta', flag: '🇲🇹' },
+  { value: 'Sweden', label: 'İsveç', flag: '🇸🇪' },
+  { value: 'Czechia', label: 'Çekya', flag: '🇨🇿' },
+  { value: 'Croatia', label: 'Hırvatistan', flag: '🇭🇷' },
+  { value: 'Bulgaria', label: 'Bulgaristan', flag: '🇧🇬' },
+  { value: 'Finland', label: 'Finlandiya', flag: '🇫🇮' },
+  { value: 'Slovenia', label: 'Slovenya', flag: '🇸🇮' },
+  { value: 'Denmark', label: 'Danimarka', flag: '🇩🇰' },
+  { value: 'Norway', label: 'Norveç', flag: '🇳🇴' },
+  { value: 'Estonia', label: 'Estonya', flag: '🇪🇪' },
+  { value: 'Lithuania', label: 'Litvanya', flag: '🇱🇹' },
+  { value: 'Luxembourg', label: 'Lüksemburg', flag: '🇱🇺' },
+  { value: 'Ukraine', label: 'Ukrayna', flag: '🇺🇦' },
+  { value: 'Latvia', label: 'Letonya', flag: '🇱🇻' }
+];
