+++ package.json
@@ -13,6 +13,7 @@
     "next": "^14.2.23",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "style-loader": "^3.3.4"
+    "style-loader": "^3.3.4",
+    "zustand": "^5.0.3"
   }
 }
